update burn  logic

+++ src/KeyBurner.sol
@@ -7,25 +7,20 @@ import {IHashFitKey} from "./IHashFitKey.sol";
 /**
  * @title Key Burner
  * @author Ibrahim
- * @notice 
+ * @notice
  */
 
-contract KeyBurner is IERC721Receiver{
-    event KeyDestroyed(address ca, uint keyId);
-    function onERC721Received(
-        address,
-        address,
-        uint256,
-        bytes calldata data
-    ) external returns (bytes4){
-        (address ca, uint keyId) = abi.decode(data,(address, uint256));
+contract KeyBurner is IERC721Receiver {
+    event KeyDestroyed(address ca, uint256 keyId);
+
+    function onERC721Received(address, address, uint256, bytes calldata data) external returns (bytes4) {
+        (address ca, uint256 keyId) = abi.decode(data, (address, uint256));
         _burn(ca, keyId);
         emit KeyDestroyed(ca, keyId);
         return IERC721Receiver.onERC721Received.selector;
     }
 
-    function _burn(address _ca, uint _keyId) internal{
+    function _burn(address _ca, uint256 _keyId) internal {
         IHashFitKey(_ca).destroyKey(_keyId);
     }
-
 }
